refactor: constant, utils 디렉토리 추가, 파일 이동
